changed file names to lowercase and adjusted the login-button to bottom
