--- src/client/components/Button/Button.style.css
@@ -1,124 +1,0 @@
-.icon{
-    font-size: 2.1875rem;
-    color: rgb(94, 97, 99);
-    margin-top: 0.2em;
-
-}
-
-p{
-    color: rgb(103, 107, 110);
-    font-size:  1em;
-    font-weight:  300;
-    line-height:  1.1em;
-    margin: 0.5em 0;
-}
-
-  .item{
-    background-color: #EBECF0;
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius: 20px;
-    box-shadow:-3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #C9CBD8;
-    display: flex;
-    color: rgba(59, 67, 73, 1);
-    flex-direction: column;
-    margin: 1em auto;
-    padding: 1em 0 ;
-    text-align:center;
-    transition:none;
-    width: 75%;
-    align-items: center;
-    outline: none;
-    
-}
-
-.item:focus-visible {
-  background-color:#EBECF0;
-  box-shadow: inset 2px 2px 8px #C9CBD8, inset -3px -3px 10px rgba(255,255,255,0.9);
-}
-.item:hover {
-    background-color:#EBECF0;
-    box-shadow: inset 2px 2px 8px #C9CBD8,
-                inset -3px -3px 10px rgba(255,255,255,0.9);
-                
-}
-.item:hover.icon{
-    color:#73ABD7;
-}
-.item:hover p{
-  color:#73ABD7;
-}
-
-
-  .itemActive {
-    background-color:#EBECF0;
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius:20px;
-    box-shadow: inset 2px 2px 8px #C9CBD8, inset -3px -3px 10px rgba(255,255,255,0.9);
-    color: rgba(59, 67, 73, 0.75);
-    flex-direction:column;
-    margin: 1em auto;
-    padding: 0.5em 0;
-    text-align:center;
-    transition: none;
-    width: 75%;
-    align-items: center;
-  }
-
- /*  .itemActive .icon,.itemActive p{
-jjj
-} */
-
-
-
-.user{
-  bottom: 0;
-  left: 1.7em;
-  position: fixed;
-  width:98px;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-  .itemActive {
-    background-color:#EBECF0;
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius:20px;
-    box-shadow: inset 2px 2px 8px #C9CBD8, inset -3px -3px 10px rgba(255,255,255,0.9);
-    color: rgba(59, 67, 73, 0.75);
-    flex-direction:column;
-    margin: 1em auto;
-    padding: 0.5em 0;
-    text-align:center;
-    transition: none;
-    width: 75%;
-    align-items: center;
-  }
-  
-  /*.itemActive .icon,.itemActive p{
-    color:#73ABD7;
-    
-}*/
-
-
-
-.user{
-  bottom: 0;
-  left: 1.7em;
-  position: fixed;
-  width:98px;
-}
-      
-
