+++ src/client/components/DashbordItems/Dashbord.style.css
@@ -13,7 +13,6 @@
   margin: 0 0 0 -0.6em;
   text-align: center;
 }
-
 .list li:hover {
   background-color: #ebecf0;
   box-shadow: inset 2px 2px 8px #c9cbd8,
