+++ src/client/components/navigation-button/navigation-button.style.css
@@ -1,0 +1,79 @@
+.icon {
+  font-size: 2.1875rem;
+  color: rgb(94, 97, 99);
+  margin-top: 0.2em;
+}
+p {
+  color: rgb(103, 107, 110);
+  font-size: 1em;
+  font-weight: 300;
+  line-height: 1.1em;
+  margin: 0.5em 0;
+}
+.item {
+  background-color: #ebecf0;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  border-radius: 20px;
+  box-shadow: -3px -3px 10px rgba(255, 255, 255, 0.8), 3px 3px 15px #c9cbd8;
+  color: rgba(59, 67, 73, 1);
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  align-items: center;
+  margin: 1em auto;
+  padding: 1em 0;
+  transition: none;
+  width: 75%;
+  outline: none;
+  position: relative;
+}
+.item:focus-visible {
+  background-color: #ebecf0;
+  box-shadow: inset 2px 2px 8px #c9cbd8,
+    inset -3px -3px 10px rgba(255, 255, 255, 0.9);
+}
+.item:hover {
+  background-color: #ebecf0;
+  box-shadow: inset 2px 2px 8px #c9cbd8,
+    inset -3px -3px 10px rgba(255, 255, 255, 0.9);
+}
+.item:hover .icon {
+  color: #73abd7;
+}
+.item:hover p {
+  color: #73abd7;
+}
+.itemActive {
+  background-color: #ebecf0;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  border-radius: 20px;
+  box-shadow: inset 2px 2px 8px #c9cbd8,
+    inset -3px -3px 10px rgba(255, 255, 255, 0.9);
+  color: rgba(59, 67, 73, 0.75);
+  flex-direction: column;
+  margin: 1em auto;
+  padding: 0.5em 0;
+  text-align: center;
+  transition: none;
+  width: 75%;
+  align-items: center;
+}
+
+.itemActive {
+  background-color: #ebecf0;
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  border-radius: 20px;
+  box-shadow: inset 2px 2px 8px #c9cbd8,
+    inset -3px -3px 10px rgba(255, 255, 255, 0.9);
+  color: rgba(59, 67, 73, 0.75);
+  flex-direction: column;
+  margin: 1em auto;
+  padding: 0.5em 0;
+  text-align: center;
+  transition: none;
+  width: 75%;
+  align-items: center;
+}
+.user {
+  margin-top: auto;
+}
